requirements.txt
spacy~=3.5
pandas
sentence-transformers
faiss-cpu
requests
tqdm
PyMuPDF
pdfplumber
